Moved functions into client library
