--- packages/types/index.ts
@@ -1,12 +1,0 @@
-export type DeepAsset = {
-  id: string;
-  tokenId: string;
-  name: string;
-  description: string;
-  image: string;
-  collection: {
-    id: string;
-    name: string;
-  };
-  address?: string;
-};
