--- packages/types/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@sni/types",
-  "version": "0.0.0",
-  "license": "GPL-3.0"
-}
